# orchestrate_bridge/openapi.template.yaml
openapi: 3.0.0
info:
  title: Repro-It Orchestrate Bridge
  description: |
    Optional HTTP bridge for running Repro-It demos via Orchestrate.
    
    This bridge allows Orchestrate to trigger Repro-It test reproduction runs.
    The main Repro-It CLI remains the source of truth and works independently.
    
    Security notes:
    - All paths are validated to prevent directory traversal
    - No secrets are exposed via this API
    - Environment variables are never printed or returned
  version: 1.0.0
servers:
  - url: https://REPLACE_ME_WITH_PUBLIC_URL
    description: Public tunnel URL (ngrok, cloudflared, etc.)
paths:
  /health:
    get:
      operationId: health_check
      summary: Health check
      description: Returns server health status
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
  /run-repro:
    post:
      operationId: run_repro_it_demo
      summary: Run Repro-It test reproduction
      description: |
        Executes Repro-It to reproduce a bug and generate a failing test.
        
        This endpoint runs the repro_it.py script with the specified bug and repo paths.
        It validates all paths for security, executes with a 120-second timeout,
        and returns detailed results including stdout, stderr, and parsed metadata.
        
        Default paths work with the included demo repository and bug examples.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                bug_path:
                  type: string
                  description: Relative path to bug JSON file
                  example: bugs/discount_double_gift_card.json
                  default: bugs/discount_double_gift_card.json
                repo_path:
                  type: string
                  description: Relative path to repository directory
                  example: demo_repo
                  default: demo_repo
      responses:
        '200':
          description: Command executed (check success field for result)
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - exit_code
                  - success
                  - judge_provider
                  - failure_summary
                  - stdout
                  - stderr
                properties:
                  ok:
                    type: boolean
                    description: Whether the API call succeeded
                    example: true
                  exit_code:
                    type: integer
                    description: Process exit code (0 = success)
                    example: 0
                  success:
                    type: boolean
                    description: Whether repro_it.py succeeded (exit_code == 0)
                    example: true
                  judge_provider:
                    type: string
                    nullable: true
                    description: AI judge provider used (e.g., "watsonx.ai", "deterministic")
                    example: watsonx.ai
                  failure_summary:
                    type: string
                    description: Summary of any failure that occurred
                    example: ""
                  stdout:
                    type: string
                    description: Standard output from repro_it.py
                    example: "Running Repro-It...\nTest generated successfully"
                  stderr:
                    type: string
                    description: Standard error from repro_it.py
                    example: ""
        '400':
          description: Invalid request (e.g., path validation failed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  exit_code:
                    type: integer
                    example: -1
                  success:
                    type: boolean
                    example: false
                  judge_provider:
                    type: string
                    nullable: true
                    example: null
                  failure_summary:
                    type: string
                    example: "Invalid bug_path: Path traversal (..) is not allowed"
                  stdout:
                    type: string
                    example: ""
                  stderr:
                    type: string
                    example: ""
        '500':
          description: Server error (e.g., timeout, unexpected exception)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  exit_code:
                    type: integer
                    example: -1
                  success:
                    type: boolean
                    example: false
                  judge_provider:
                    type: string
                    nullable: true
                    example: null
                  failure_summary:
                    type: string
                    example: "Command timed out after 120 seconds"
                  stdout:
                    type: string
                    example: ""
                  stderr:
                    type: string
                    example: "Timeout"
  /run-repro-from-text:
    post:
      operationId: reproduce_bug_from_text
      summary: Run Repro-It from plain-text bug report
      description: |
        Accepts a plain English bug report, converts it to JSON format,
        executes Repro-It to reproduce the bug, and generates a failing regression test.
        
        **Important:** This endpoint generates a failing test to reproduce the bug.
        It does NOT fix bugs or provide solutions. The generated test serves as
        a regression test that developers can use to verify their fix.
        
        **Required Information:**
        The bug report must include all three:
        1. Expected behavior (e.g., "expected", "should", "supposed to")
        2. Observed behavior (e.g., "actual", "got", "returns", "shows", "error")
        3. Feature/function context (e.g., "discount", "DOUBLE", "parse_price", "checkout")
        
        Vague reports like "My app crashes sometimes" will be rejected with HTTP 400.
        
        **Valid Example:**
        "Hey, our DOUBLE discount code is not working on gift card orders. A customer
        expected to save 20% on her $100 gift card. Instead the total returns $100
        with no discount applied."
        
        The converted bug JSON is saved to bugs/from_orchestrate.json.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Plain-text bug report (required, non-empty)
                  example: "QA says discount code DOUBLE is broken on gift cards. Expected 20% off but checkout shows full price $100."
                bug_id:
                  type: string
                  description: Bug ID for tracking
                  example: BUG-orch-001
                  default: BUG-orch-001
                severity:
                  type: string
                  description: Bug severity level
                  enum: [low, medium, high, critical]
                  example: medium
                  default: medium
                repo_path:
                  type: string
                  description: Relative path to repository directory
                  example: demo_repo
                  default: demo_repo
      responses:
        '200':
          description: Command executed (check success field for result)
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - exit_code
                  - success
                  - judge_provider
                  - failure_summary
                  - converted_bug
                  - stdout
                  - stderr
                properties:
                  ok:
                    type: boolean
                    description: Whether the API call succeeded
                    example: true
                  exit_code:
                    type: integer
                    description: Process exit code (0 = success)
                    example: 0
                  success:
                    type: boolean
                    description: Whether repro_it.py succeeded
                    example: true
                  judge_provider:
                    type: string
                    nullable: true
                    description: AI judge provider used
                    example: watsonx.ai
                  failure_summary:
                    type: string
                    description: Summary of test failure
                    example: "E   AssertionError: Expected 80.0, got 100.0"
                  converted_bug:
                    type: object
                    description: The bug JSON created from plain text
                    properties:
                      id:
                        type: string
                        example: BUG-orch-001
                      title:
                        type: string
                        example: "QA says discount code DOUBLE is broken on gift cards."
                      description:
                        type: string
                        example: "QA says discount code DOUBLE is broken on gift cards. Expected 20% off but checkout shows full price $100."
                      reporter:
                        type: string
                        example: orchestrate
                      severity:
                        type: string
                        example: medium
                      expected:
                        type: string
                        example: "Expected 20% off"
                      observed:
                        type: string
                        example: "checkout shows full price $100"
                  stdout:
                    type: string
                    description: Standard output from repro_it.py
                    example: "Running Repro-It...\nTest generated successfully"
                  stderr:
                    type: string
                    description: Standard error from repro_it.py
                    example: ""
        '400':
          description: Invalid request (missing text, too vague, or invalid path)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Bug report is too vague to reproduce"
                  details:
                    type: string
                    example: "Please include expected behavior, actual behavior, and the feature or function involved."
        '500':
          description: Server error (timeout or unexpected exception)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  exit_code:
                    type: integer
                    example: -1
                  success:
                    type: boolean
                    example: false
                  judge_provider:
                    type: string
                    nullable: true
                    example: null
                  failure_summary:
                    type: string
                    example: "Command timed out after 120 seconds"
                  converted_bug:
                    type: object
                    description: Partial bug JSON if conversion succeeded
                    example: {}
                  stdout:
                    type: string
                    example: ""
                  stderr:
                    type: string
                    example: "Timeout"

# Made with Bob